rename files for clarity
also made the naming convention consistent (dashes not underscore)
