+++ scripts/validate-translations.js
@@ -19,7 +19,7 @@ if (args.length < 1) {
 import fs from 'fs';
 import path from 'path';
 import async from 'async';
-import {validateTranslations} from './tx_util.js';
+import {validateTranslations} from './tx-util.js';
 import locales from '../src/supported-locales.js';
 
 // Globals
